gradio and Hugging face spaces/requirements.txt
ffmpeg-python==0.2.0
gradio==5.13.2
librosa==0.10.2.post1
noisereduce==3.0.3
numpy==2.1.3
sentencepiece==0.2.0
torch==2.5.1
transformers==4.48.1